SG-13994 u-boot: use our own fw_env.config

+++ recipes-bsp/u-boot/u-boot-fw-utils_2019.07.bb
@@ -7,6 +7,10 @@ INSANE_SKIP_${PN} = "already-stripped"
 EXTRA_OEMAKE_class-target = 'CROSS_COMPILE=${TARGET_PREFIX} CC="${CC} ${CFLAGS} ${LDFLAGS}" HOSTCC="${BUILD_CC} ${BUILD_CFLAGS} ${BUILD_LDFLAGS}" V=1'
 EXTRA_OEMAKE_class-cross = 'HOSTCC="${CC} ${CFLAGS} ${LDFLAGS}" V=1'
 
+SRC_URI_append += " \
+	file://fw_env.config \
+"
+
 inherit uboot-config
 
 do_compile () {
@@ -19,7 +23,11 @@ do_install () {
 	install -d ${D}${sysconfdir}
 	install -m 755 ${S}/tools/env/fw_printenv ${D}${base_sbindir}/fw_printenv
 	install -m 755 ${S}/tools/env/fw_printenv ${D}${base_sbindir}/fw_setenv
-	install -m 0644 ${S}/tools/env/fw_env.config ${D}${sysconfdir}/fw_env.config
+	if [ -e ${WORKDIR}/fw_env.config ] ; then
+		install -m 0644 ${WORKDIR}/fw_env.config ${D}${sysconfdir}/fw_env.config
+	else
+		install -m 0644 ${S}/tools/env/fw_env.config ${D}${sysconfdir}/fw_env.config
+	fi
 }
 
 do_install_class-cross () {
